requirements.txt
discord.py
openai
pdfplumber
python-dotenv